Problems with the functions makeRelayOn(i); and makeRelayOff(i);
To fix! - The upper functions correctly retrieve information about the relays affected by the schedule from the EEPROM, but when they try to write to the eeprom the new relay state, HAL_GetTick(); stop working. When there is no relay in the schedule whose state is supposed to change, the functions work correctly.